Added github info message to config.yml in case the config gets messed up by the player or is missing the comment. (probably need to implement a custom config updater because the default one success majorly and is removing all my comments and explanations...)
Also added permAndArgsCheck method that already contains the perms check
from the previous commit. That way it's harder to forget checking the
amount of arguments (no more out of array exceptions *yay*). This seems
to be the most elegant solution since I want to keep my individual
command-execution method as lean as possible.